Fix the compilation error in test code.

+++ test/com/jwetherell/algorithms/mathematics/test/CoprimesTest.java
@@ -16,6 +16,6 @@ public class CoprimesTest {
         final List<Long> expected = Arrays.asList(1L, 16L, 32L, 164L, 1194891264L);
 
         for(int i = 0; i < args.size(); i++)
-            assertEquals(expected.get(i), Coprimes.getNumberOfCoprimes(args.get(i)));
+            assertEquals(expected.get(i).longValue(), Coprimes.getNumberOfCoprimes(args.get(i)));
     }
 }
